# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервер для интернет-магазина продуктов FoodStore",
        "title": "FoodStore API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/protected/moderator/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Одобряет предложение продавца на модерации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderator"
                ],
                "summary": "Одобрить предложение",
                "parameters": [
                    {
                        "description": "Данные одобрения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OfferModerationAnswerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OfferModerationAnswerOutput"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Предложение не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/moderator/offers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список предложений продавцов, ожидающих модерацию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderator"
                ],
                "summary": "Получить предложения на модерацию",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.OfferModerationOutput"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/moderator/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех существующих продуктов для модератора",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderator"
                ],
                "summary": "Получить существующие продукты",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CategoryIDOutput"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/moderator/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отклоняет предложение продавца на модерации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderator"
                ],
                "summary": "Отклонить предложение",
                "parameters": [
                    {
                        "description": "Данные отклонения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OfferModerationAnswerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OfferModerationAnswerOutput"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Предложение не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
...
        "/recovery-password/reset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Сбрасывает пароль пользователя (требуется recovery токен)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Сбросить пароль",
                "parameters": [
                    {
                        "description": "Новый пароль",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResetPasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResetPasswordOutput"
                        }
                    },
                    "400": {
                        "description": "Пустые поля",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/recovery/send-code": {
            "post": {
                "description": "Отправляет код восстановления пароля на email пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Отправить код восстановления",
                "parameters": [
                    {
                        "description": "Данные для отправки кода",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendCodeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SendCodeOutput"
                        }
                    },
                    "400": {
                        "description": "Пустые поля",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/recovery/verify-code": {
            "post": {
                "description": "Проверяет код восстановления и возвращает recovery токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Проверить код восстановления",
                "parameters": [
                    {
                        "description": "Данные для проверки кода",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VerifyCodeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.VerifyCodeOutput"
                        }
                    },
                    "400": {
                        "description": "Пустые поля или неверный код",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/refresh-access/token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет access токен по refresh токену",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Обновить access токен",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RefreshAccessTokenOutput"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CategoryIDOutput": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sub_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubCategoryIDOutput"
                    }
                }
            }
        },
...
        "dto.ErrorOutput": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.OfferModerationAnswerInput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "dto.OfferModerationAnswerOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "dto.OfferModerationOutput": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                },
                "sub_category_id": {
                    "type": "integer"
                },
                "sub_category_name": {
                    "type": "string"
                }
            }
        },
        "dto.ProductIDOutput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.RefreshAccessTokenOutput": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResetPasswordInput": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "dto.ResetPasswordOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.SendCodeInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "dto.SendCodeOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.SubCategoryIDOutput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductIDOutput"
                    }
                },
                "sub_category_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.VerifyCodeInput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "dto.VerifyCodeOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "recovery_token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}