# internal/delivery/http/handler/recovery_handler.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/yakupovdev/FoodStore/internal/delivery/http/dto"
	"github.com/yakupovdev/FoodStore/internal/usecase"
)

type RecoveryHandler struct {
	uc *usecase.RecoveryUsecase
}

func NewRecoveryHandler(uc *usecase.RecoveryUsecase) *RecoveryHandler {
	return &RecoveryHandler{uc: uc}
}

// ResetUserPassword godoc
// @Summary Сбросить пароль
// @Description Сбрасывает пароль пользователя (требуется recovery токен)
// @Tags Recovery
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param input body dto.ResetPasswordInput true "Новый пароль"
// @Success 200 {object} dto.ResetPasswordOutput
// @Failure 400 {object} dto.ErrorOutput "Пустые поля"
// @Failure 404 {object} dto.ErrorOutput "Пользователь не найден"
// @Failure 500 {object} dto.ErrorOutput
// @Router /recovery-password/reset [post]
func (h *RecoveryHandler) ResetUserPassword(ctx *gin.Context) {
	var req dto.ResetPasswordInput

	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(ErrInvalidJSON.Status, ErrInvalidJSON.Response)
		return
	}

	if req.NewPassword == "" {
		ctx.JSON(ErrEmptyFields.Status, ErrEmptyFields.Response)
		return
	}

	uid, ok := extractUserID(ctx)
	if !ok {
		ctx.JSON(ErrUserNotFound.Status, ErrUserNotFound.Response)
		return
	}
	req.UserID = uid

	err := h.uc.ResetPassword(ctx.Request.Context(), req)
	if err != nil {
		ctx.JSON(ErrInternal.Status, ErrInternal.Response)
		return
	}

	ctx.JSON(http.StatusOK, dto.ResetPasswordOutput{Message: "Password reset"})
}